requirements.txt
et_xmlfile==2.0.0
ffmpeg-python==0.2.0
future==1.0.0
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
